cambios solicitados + transacciones
